pyproject.toml
[tool.poetry]
name = "math-gorilla"
version = "0.1.0"
description = ""
authors = ["pizzacat83 <17941141+pizzacat83@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
